Fixes #270: Parse error: syntax error, unexpected T_VARIABLE in \htdocs\include\functions.php on line 1914.

+++ htdocs/include/functions.php
@@ -1847,7 +1847,7 @@ function icms_escapeValue($value, $quotes = true)
 			$dd = opendir($d);
 			while($file = readdir($dd))
 			{
-		 		if(is_file($d.$file) && ($file != 'index.html' && $file != 'php.ini' && $file != '.htaccess' $file != '.svn'))
+		 		if(is_file($d.$file) && ($file != 'index.html' && $file != 'php.ini' && $file != '.htaccess' && $file != '.svn'))
 				{
 		  			unlink($d.$file);
 				}
